pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mall-admin-api"
version = "0.1.0"
description = "商城管理系统后端API"
requires-python = ">=3.8"
dependencies = [
    "fastapi",
    "uvicorn",
    "sqlalchemy",
    "pydantic",
    "pydantic-settings",
    "python-jose[cryptography]",
    "passlib[bcrypt]",
    "python-multipart",
    "alembic",
    "pymysql",
    "aioredis",
    "celery",
    "python-dotenv",
] 